Separate Choreographer callbacks between sessions
This change completely substitutes default (Android)
Choreographer callbacks management with custom
layoutlib behavior. In order to run callbacks
selectively (specific to RenderSession) we have
to change internal Choreographer callback queue,
in that case it seems easier to completely substitute
it with custom a queue. In addition to that we have
a more predictible behavior as Choreographer doFrame
execution has a number of side effects.

Bug: 158750833
Fixes: 159805328
Test: all render tests should pass
Change-Id: I1d0eec909f75f28e5512dbea9a6d964bdd0a9294