Fix custom Cleaner implementation
Calling clean() on a layoutlib Cleaner objet was triggering the clean
action, but this was bypassing the phantom reference registered with the
JVM cleaner. That meant that the JVM could still decide to again trigger
the clean action on the same object at a later time, with unpredictable
behaviour.
With this change, the clean() method delegates to the reference created
by the JVM cleaner, ensuring that the cleaning action is only called
once.

Bug: N/A
Test: N/A
Change-Id: I3193d7b144bbd0d9491c574f7bc47ffcb2ff1fb5
(cherry picked from commit 10edfa99cff9d03bfd3755852ae11fa760bf1df2)