Support AnimationHandler in interactive preview
The animation handler relies on the choreographer for timing. This
implements Choreographer.getFrameTime to directly use the system time to
be in sync with the other callbacks.
This also makes sure to clear all the AnimationHandler callbacks when
disposing the RenderSession to avoid potential leaks.

Bug: 189093606
Test: checked manually that this enables support for software-rendered
ripples.

Change-Id: I84f10f80e072612ac4d7b9e43cf6f9623531c788
(cherry picked from commit 8ef4e9f7c0f8dd02516383ca34d17373551cefb9)