Update AssetManager_Delegate
AssetManager.nativeCreateIdmapsForStaticOverlaysTargetingAndroid has
been deleted, so we remove its delegate.
AssetManager.nativeThemeDestroy has been deleted, and
AssetManager.nativeGetThemeFreeFunction has been created instead. We
delete the delegate of the first one, and implement a new delegate for
the second one which is responsible for removing references to
Resources_Theme_Delegate objects.

Bug: N/A
Test: re-enable TestDelegates.testMethodDelegates that would have caught
the problem earlier.

Change-Id: Ib28160b5e0f4d68426dd8672bcb74c417e78b638