Push resource lifecycle to the session scope
Before this CL, the Resources would be initialiazed in the
RenderAction.setUp and disposed in tearDown. These are meant to be done
for every render/callback execution and be short lived.
Some features, like live literals, rely on the resources being
available until the session is fully disposed. This change allows
resources to not be released early.

Test: Existing tests pass
Bug: 202548478
Change-Id: Id07134b96b998f2e525062c79c39eafc47b7b1b6