Create script to split universal binaries for layoutlib
The Android build system builds universal binaries of its native
libraries, containing x86_64 and arm64 versions. For layoutlib
we want to split the architectures into individual files that
can be shipped only for the relevant platform, in order to save
space.
This script, to be run after the build step, does the architecture
splitting and adds the file into the DIST directory so that they
can be downloaded from ab/.

Bug: N/A
Test: N/A
Change-Id: Iaf585c166c6f8f476b4c144508b5d4fc8f4b15ce