Better errors when failing to find custom preference
When inflating preferences, if a custom class is not found, log the
error to better inform the user what the problem is. Right now, this
causes an error later on which is confusing as the root cause is not
clear.

Fixes: 37132984
Test: checked from Studio
Change-Id: I81e521ac1f325b473fdc6fdc251f4ff78db73324
(cherry picked from commit 2c7c59d037e0123967d4a859c1fceadcb8096b7e)