Fix method replacer used in AdaptiveIconDrawable
We want to replace a call to a non-static method by a call to a static
method. In order to do this, we need to use as the first argument of our
static replacement the object on which the non-static method is
originally called. That is because this object is loaded on the stack,
and we need to use it for the rest of the bytecode to be valid.

Bug: 74598470
Test: layoutlib tests
Change-Id: Ibf39080ee483ad2f6b99415848156f68622a3162