Resources_Delegate.getColor should access statelists
When accessing a color resource through Resources.getColor, if the
resource resolves to a color value, then the method returns that color.
But if it resolves to a color state list, it returns the default color
for that state list. Resources_Delegate.getColor should do the same
instead of failing for state lists.

Bug: N/A
Test: layoutlib tests
Change-Id: I9bcf12d16e47bac3b36d9e036ed6fb2f615d70e8