Create a delegating Application Context
Animations in Compose need to store the Application Context which is
currently the BridgeContext in Layoutlib. That results in a leak of the
BridgeContext which is a problem. Instead this creates an Application
Context that delegates back to the Bridge Context through a weak
reference.

Bug: 243418281
Test: To be checked in Studio
Change-Id: Ia35dff0f7b545c497839191e24c67aa99ac9271b
(cherry picked from commit 3445b0d6f094695c390214cee2d2424f0caf0654)
Merged-In: Ia35dff0f7b545c497839191e24c67aa99ac9271b