More robust ProcessRecord routing for resultTo.
To deliver a "resultTo" listener for a broadcast intent, we need to
know about the ProcessRecord we're returning the result to, so that
we can potentially unfreeze them and sequence the response through
their main thread.

Also offer to lookup the "system_server" process when the caller is
the system_server PID and UID; this smoothly handles dozens of
situations where the system sends broadcasts without having easy
access to its own ProcessRecord.

Bug: 251486235
Test: manual
Change-Id: Ic3958d1ca376a8bf04e144d883373226ba7fbd7c