Refactor parts of layoutlib/bridge/tests into a bridge_client.
Currently layoutlib/bridge has tests that demonstrate how to use the
bridge to inflate layouts and render views. This code is quite useful
and is being refactored into a library:
  bridge_client (layoutlib-bridge-client).
Part of the work was already done on 260244331, but most of the code
review comments were not adhered to before submitting.

Bug: 260244331
Test: run_tests.sh for regression.
Change-Id: Ic16a1f63e34798c9124487be5372448a5a2a8f4e