Fix IntelliJ structure of the layoutlib project
This creates an IntelliJ module for each Soong library, and adds the
corresponding dependencies. Dependencies for IntelliJ modules are not
transitive which is why there are more dependencies in the .iml file
than the corresponding .bp file.
In addition, this cleans up dependencies, IntelliJ libraries, and some
broken package names.

Bug: 260244331
Test: layoutlib tests from IntelliJ and from running run_tests.sh
Change-Id: Ia75554c9cb49e7099be06fe0f112278f945094c8