Replace calls to Application.getResources in Html.java
In Html.java, there are several calls trying to get the resources from
the application. It first calls ActivityThread.currentApplication(), and
then calls getResources on the value returned by the previous call.
In layoutlib, ActivityThread.currentApplication() returns null, so the
next call to getResources triggers an NPE.
To avoid this, we replace the call to getResources by a static call to
Application_Delegate.getResources which returns the resources without
the need for a non-null Application object.

Bug: 262773698
Test: test added
Change-Id: If22991ea190de06388c6fee0794e6b271f09c4b7