Correctly trim quotes and/or spaces from string resources
Android string resources behave in the following way:
If string resource values are surrounded by quotes, the quotes need to
be removed, but all spaces inside the quotes are kept as is.
Otherwise, spaces at the beginning or the end of the value have to be
removed.

This ensures that this behaviour is applied to TextResourceValue.

Bug: 266186515
Test: N/A
Change-Id: I48a387bdf785daf5f3ae408a7b33ead598cc7d83
(cherry picked from commit e69fee9011ab791cd118f47843e869a1dca60e6c)
Merged-In: I48a387bdf785daf5f3ae408a7b33ead598cc7d83