Modify scaling for RenderDrawable
When rendering a drawable, this first renders it full size and then uses
bitmap scaling to potentially downscale it to the required size. This
ensures that the full drawable is visable in the end, as trying to
render on a smaller image at first can sometimes result in only a small
part of the original drawable being rendered.
In addition, this cleans up the conversion between Bitmap and
BufferedImage to avoid unnecessary allocations.

Fixes: 173833413
Test: Tested from Studio
Change-Id: I49e0b09144f35aeecbfe7071a4a3e91c63f32590