Correctly handle string resources with HTML and quotes
For string resources, contiguous whitespace characters and new lines are
collapsed into one space except when the string is surrounded by quotes.
However, that behaviour is altered when there are HTML tags in the
string. In this case, as soon as the first tag is encountered,
whitespaces are always collapsed even inside quotes.
This also makes sure the same behaviour is applied to strings coming
from a typed array or strings directly accessed through resources.

Bug: 246401259
Test: Test added
Change-Id: I76b65a8f1c90ce86ceb1318e763613311a2a5128