Fix ripple animation from the framework
This uses the Choreographer to sync the time for animations to display.
This ensures that different render tasks can have their own time despite
the fact that the RenderThread only has one time that can only be
increased.
This also simplifies the rendering code by reusing more of the
underlying platform code, especially ThreadedRenderer and Choreographer.

Bug: 290775877
Test: layoutlib tests
Change-Id: I9046c26da2c3e9a937c533a1640a65802cec9bf1