Use ApplicationContext to create BridgeContentResolver
BridgeContentResolver holds a reference to the context that is used to
create it. This means that if user/library code stores a content
resolver, that can lead to a leak of the BridgeContext. To prevent this,
we create BridgeContentResolver with the ApplicationContext, which only
holds a weak reference to the BridgeContext.

Bug: 290990640
Test: N/A
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:044f194cf028afac353634b618a486110d9feb52)
Merged-In: Ib5a5429d2bb11a75dc5ecfdf20effc98b797a2cc
Change-Id: Ib5a5429d2bb11a75dc5ecfdf20effc98b797a2cc