Switch NativeAllocationRegistry.applyFreeFunction to a MethodReplacer
Previously, NativeAllocationRegistry.applyFreeFunction was a delegate
method, and the native code had to bind to a separate method on
NativeAllocationRegistry_Delegate. Update it to use MethodReplacer
('nativeApplyFreeFunction'), which makes it possible to use the
unmodified NativeAllocationRegistry JNI code. This is the same approach
taken by Path.nInit in LayoutLib.

This makes it possible to use libcore_util_NativeAllocationRegistry.cpp
in places where LayoutLib delegates are not available (e.g. RNG).

Bug: 315497747

Test: layoutlib-tests and verified that applyFreeFunction is called
Change-Id: I964ede96fa7b596c7b866b4e0517496484c1843d