Bypass call to Application.getApplicationInfo
In LineBreakConfig, there is a call to
ActivityThread.currentApplication, which returns null for Layoutlib. On
this is then called getApplicationInfo, which fails with a NPE.
This replaces that last call with a call to a static delegate, which
accepts a null application as argument, and returns the current
ApplicationInfo from the BridgeContext.

Bug: 74062470
Test: layoutlib tests
Change-Id: I3e98a36648c4e12029e6cd95fd9af66d09c2ddcb