Add edge to edge handling to sys UI
This sets up the insets and transparency settings to the layoutlib
version of sys UI. Edge to edge is on by default in Android V so this
enables it in layoutlib as well.
Instead of having the system bars in the same layout as the content,
this sets them up in their own fake window and provides the insets to
the window manager.

Bug: 74062470
Test: tests updated
Change-Id: I08329bc797e0c940b38c90d52942dc3627dd4d23