Create layoutlib_jni for setting up native layoutlib
This creates a new JNI library that will serve as the entry point for
layoutlib. Upon loading, it performs a number of tasks specific to
layoutlib:
- Registering JNI methods for LayoutlibRenderer
- Setting up logger/aborter to call in native code
- Calling Bridge.setSystemProperties that sets up a number of properties
  needed by layoutlib
- Setting up the keyboard mappings for input support
This library then calls on the host runtime of libandroid_runtime to
set up the entire native graphics code from the Android platform.

Flag: NONE layoutlib for host
Bug: 322360037
Test: render tests once everything is set up
Change-Id: If68b62cabef445565532689a64a3a2781335a471