Add screen wake lock listener to PowerManager
This is needed for DeviceStateProvider on foldable
devices to force using tent mode while there is
a wakelock keeping the screen 'on'.

This is also could be used in the future for
analytics reporting to understand how often we have
screen wakelock available when switching
between displays.

This is hidden API to be used by system_server's
DeviceStateProvider or SystemUI.
The method will be called under a trunk stable flag.

Test: atest PowerManagerServiceTest
Bug: 363174979
Flag: EXEMPT bugfix
Change-Id: I5f4b6ece52a0689c2c1441d3869bdeaadb5dac51