Fix memory leak in layoutlib buffer creation
The BufferItem object is used to acquire a GraphicBuffer object from
the BufferQueue. Once that is done, there is no use for the BufferItem
anymore, and it needs to be deleted so that it does not leak the
GraphicBuffer since it holds a reference to it.

Flag: NONE host-only change
Bug: 369654164
Test: checked manually
Change-Id: I21701ae1f25e4438f1dbf98a0f2323d4211d4196