Do not create HostRuntime for layoutlib
On Windows, we need to load libandroid_runtime.dll first as it is a
dependency of layoutlib_jni. But doing so calls the Jni_OnLoad method of
libandroid_runtime, which we do not want. Setting the
use_base_native_hostruntime property to false ensures that we do not
execute the content of that method.

Flag: NONE host-only change
Bug: 340885449
Test: N/A
Change-Id: Ia65dc86fee26826a3ada45359dd0ca2ccf54edee