Do not add unnecessary AndroidX classes to layoutlib jar
When looking for classes inheriting from android View or Fragment, we
should not consider AndroidX classes. This ensures that those are
filtered out while keeping in the AndroidX classes that are dependencies
of other included classes.
In addition, this applies jarjar to androidx and org.jetbrains packages
to avoid possible classpath conflicts when using inside Android Studio.

Flag: NONE host-only change
Bug: N/A
Test: layoutlib tests
Change-Id: Iad940c214489131843899fc045edaeb5b9f99fa5