Fix jarjar-ing of classes in layoutlib
Layoutlib builds use jarjar to rename classes from its dependencies that
could collide with classes in Android Studio. However, it should not
refactor the custom layoutlib code itself.
Thus this makes the jarjar-ing more specific to the libraries where it
is needed instead of doing it at the top-level.

Flag: NONE host-only change
Bug: 389578481
Test: to be tested from Studio
Change-Id: I6312f0a16b98fd91c6aff7657d95518c06fa7033