Clear Bitmap caches as part of the RenderAction release
Bitmap caches end up holding on to the native memory associated with
Bitmap objects, which can be quite large. This is particularly true if
the cahce is using soft references as they will only get cleared if the
Java heap is under pressure, independently of the amount of native
memory used.
To avoid holding on to large amount of native memory, this clears the
caches after each render action. Caches can be re-enabled through a
render flag.

Flag: NONE host-only change
Bug: 388141123
Bug: 396183478
Test: N/A
Change-Id: Ic5974545397e65861154c93073baaea8e0c778af